+++ app/models/move_tag.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+# Relates a move with its tags
+class MoveTag < ApplicationRecord
+  # Relationships
+  belongs_to :move
+  belongs_to :tag
+end