+++ test/controllers/moves_controller_test.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'test_helper'
 
 # Tests to assure moves can be viewed, created, and modified
@@ -17,9 +19,13 @@ class MovesControllerTest < ActionDispatch::IntegrationTest
   end
 
   test 'should create move' do
-    assert_difference('Move.count') do
-      post moves_url, params: { move: { description: @move.description, name: @move.name, video_url: @move.video_url } }
+    assert_difference('Tag.count', 1) do
+      assert_difference('Move.count') do
+        post moves_url, params: { move: { name: @move.name, description: @move.description, video_url: @move.video_url, tag_tokens: 'wrap, cross body' } }
+      end
     end
+
+    assert_equal 2, Move.last.tags.count
 
     assert_redirected_to move_path(Move.last)
   end
@@ -35,7 +41,7 @@ class MovesControllerTest < ActionDispatch::IntegrationTest
   end
 
   test 'should update move' do
-    patch move_url(@move), params: { move: { description: @move.description, name: @move.name, video_url: @move.video_url } }
+    patch move_url(@move), params: { move: { name: @move.name, description: @move.description, video_url: @move.video_url, tag_tokens: '' } }
     assert_redirected_to move_path(@move)
   end