+++ test/models/move_tag_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class MoveTagTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end