+++ CHANGELOG.md
@@ -4,3 +4,6 @@
   - Initial framework using Rails 5.0.0.beta2
 - **Moves Added**
   - Added ability to create, view, modify, and search through moves
+- **Tags Added**
+  - Moves can be tagged
+  - Tags can be used to search for similar moves